+++ client/DotaAsker/User.h
@@ -11,7 +11,7 @@
 
 @interface User : NSObject
 
-@property (assign, nonatomic) NSInteger ID;
+@property (assign, nonatomic) unsigned long long ID;
 @property (assign, nonatomic) NSInteger MMR;//Текущий рейтинг среди всех игроков
 @property (assign, nonatomic) float KDA;//Отношение правильно отвеченных вопросов на неправильно отвеченные.
 @property (assign, nonatomic) float GPM;//Насколько быстро вы отвечаете на вопросы
